requirements-dev.txt
Django==3.0.8
djangorestframework==3.11.0
dj-database-url==0.5.0
python-decouple==3.3
dj-static==0.0.6
Pillow==7.2.0